chore: change chain retry history list
