+++ packages/maskbook/src/plugins/RedPacket/UI/RedPacketDialog.tsx
@@ -7,7 +7,7 @@ import { RedPacketJSONPayload, DialogTabs } from '../types'
 import { editActivatedPostMetadata } from '../../../protocols/typed-message/global-state'
 import { RedPacketMetaKey } from '../constants'
 import { RedPacketForm } from './RedPacketForm'
-import { RedPacketBacklogList } from './RedPacketList'
+import { RedPacketHistoryList } from './RedPacketHistoryList'
 import { InjectedDialog } from '../../../components/shared/InjectedDialog'
 
 interface RedPacketDialogProps extends withClasses<never> {
@@ -49,7 +49,7 @@ export default function RedPacketDialog(props: RedPacketDialogProps) {
             },
             {
                 label: t('plugin_red_packet_select_existing'),
-                children: <RedPacketBacklogList onSelect={onCreateOrSelect} onClose={onClose} />,
+                children: <RedPacketHistoryList onSelect={onCreateOrSelect} onClose={onClose} />,
                 sx: { p: 0 },
             },
         ],
