+++ packages/maskbook/src/plugins/RedPacket/apis/index.ts
@@ -9,8 +9,7 @@ import type {
     RedPacketSubgraphInMask,
     RedPacketHistory,
 } from '../types'
-
-import { EthereumNetwork, EthereumTokenType } from '../../../web3/types'
+import { EthereumNetwork, EthereumTokenType, ChainId } from '../../../web3/types'
 
 const redPacketBasicKeys = [
     'contract_address',
@@ -86,8 +85,8 @@ export async function getRedPacketTxid(rpid: string) {
     return redPackets[0] ? redPackets[0].txid : null
 }
 
-export async function getRedPacketHistory(address: string) {
-    const response = await fetch(getConstant(RED_PACKET_CONSTANTS, 'SUBGRAPH_URL', await getChainId()), {
+export async function getRedPacketHistory(address: string, chainId: ChainId) {
+    const response = await fetch(getConstant(RED_PACKET_CONSTANTS, 'SUBGRAPH_URL', chainId), {
         method: 'POST',
         mode: 'cors',
         body: JSON.stringify({
