+++ packages/maskbook/src/plugins/RedPacket/hooks/useRedPacketHistory.ts
@@ -1,9 +1,10 @@
 import { useAsyncRetry } from 'react-use'
 import { RedPacketRPC } from '../messages'
+import type { ChainId } from '../../../web3/types'
 
-export function useRedPacketHistory(address: string) {
+export function useRedPacketHistory(address: string, chainId: ChainId) {
     return useAsyncRetry(async () => {
-        const payloads = await RedPacketRPC.getRedPacketHistoryWithPassword(address)
+        const payloads = await RedPacketRPC.getRedPacketHistoryWithPassword(address, chainId)
         return payloads
     }, [address])
 }
